fix: refactor ProcessVariablesMap class and Jackson module packages
