+++ activiti-core/activiti-api-impl/activiti-api-process-model-impl/src/test/java/org/activiti/api/runtime/test/model/impl/ProcessVariableValuesMapSerializerTest.java
@@ -1,4 +1,4 @@
-package org.activiti.api.runtime.test.conf.impl;
+package org.activiti.api.runtime.test.model.impl;
 
 import org.junit.jupiter.api.Test;
